% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian}
\alias{italian}
\title{italian resturants data}
\format{A data frame with 168 rows and 7 variables:
\describe{
  \item{price}{price, in US dollars for a meal for two}
  \item{east}{is it lokated on the east of 5the avenue yes(1) or no(0)}
  \item{resturant}{name of resturant}
  \item{food}{ranking 0-30}
}}
\source{
http://gattonweb.uky.edu/sheather/book/docs/datasets/nyc.csv
}
\usage{
italian
}
\description{
a data set of italina resutrents and their price, food, decore
}
\examples{
  # calculate the mean price
  mean(italian$price)
  if (require(dplyr)) {
    dplyr::filter(italian, price >30)
 }

}
\keyword{datasets}
