% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojifunc.R
\name{print.ranawg}
\alias{print.ranawg}
\title{print a an emoji with a data frame.}
\usage{
\method{print}{ranawg}(x, ...)
}
\arguments{
\item{x}{A data frame.}

\item{...}{other parameters}
}
\description{
print a an emoji with a data frame.
}
\examples{
print.ranawg(mtcars)
}
